feat: update btc_broadcastTX test

+++ crates/rooch-relayer/src/actor/bitcoin_client_proxy.rs
@@ -50,7 +50,7 @@ impl BitcoinClientProxy {
         &self,
         hex: String,
         maxfeerate: Option<f64>,
-        maxburnamount: Option<u64>,
+        maxburnamount: Option<f64>,
     ) -> Result<Txid> {
         self.actor
             .send(BroadcastTransactionMessage {
